Fixed bug where all game sessions were assigned the same track order
Upon getting the tracks 'randomly' from the DB, the queryset would essentially keep the original DB-ordering, as the queryset doesn't hit the database until it is 'used'. Due to this, the fix was to convert the queryset to a list (actually hit the DB and get objects), and then shuffle the list.